backend/requirements.txt
fastapi>=0.115.0
slowapi>=0.1.9
email-validator>=2.0.0
stripe>=7.0.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
bcrypt>=4.0.0
PyJWT>=2.8.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
